+++ src/Termonad/Term.hs
@@ -88,16 +88,17 @@ import GI.Vte
   , onTerminalWindowTitleChanged
   , terminalGetWindowTitle
   , terminalNew
+  , terminalSetBoldIsBright
   , terminalSetCursorBlinkMode
   , terminalSetFont
   , terminalSetScrollbackLines
-  , terminalSpawnSync
   , terminalSetWordCharExceptions
-  , terminalSetBoldIsBright
+  , terminalSpawnSync
   )
 import System.Directory (getSymbolicLinkTarget)
 import System.Environment (lookupEnv)
 
+import Termonad.Gtk (terminalSetEnableSixelIfExists)
 import Termonad.Lenses
   ( lensConfirmExit
   , lensOptions
@@ -114,7 +115,7 @@ import Termonad.Lenses
   )
 import Termonad.Types
   ( ConfigHooks(createTermHook)
-  , ConfigOptions(scrollbackLen, wordCharExceptions, cursorBlinkMode, boldIsBright)
+  , ConfigOptions(scrollbackLen, wordCharExceptions, cursorBlinkMode, boldIsBright, enableSixel)
   , ShowScrollbar(..)
   , ShowTabBar(..)
   , TMConfig(hooks, options)
@@ -360,6 +361,7 @@ createAndInitVteTerm tmStateFontDesc curOpts = do
   terminalSetScrollbackLines vteTerm (fromIntegral (scrollbackLen curOpts))
   terminalSetCursorBlinkMode vteTerm (cursorBlinkMode curOpts)
   terminalSetBoldIsBright vteTerm (boldIsBright curOpts)
+  terminalSetEnableSixelIfExists vteTerm (enableSixel curOpts)
   widgetShow vteTerm
   pure vteTerm