+++ src/Termonad/Types.hs
@@ -408,6 +408,11 @@ data ConfigOptions = ConfigOptions
     -- If 'False', then bold can be applied separately to colors from both the
     -- 'Termonad.Config.Colour.BasicPalatte' and
     -- 'Termonad.Config.Colour.ExtendedPalatte'.
+  , enableSixel :: !Bool
+    -- ^ Enable SIXEL to draw graphics in a terminal.
+    --
+    -- Note that SIXEL may not be fully supported in VTE. Follow
+    -- <https://gitlab.gnome.org/GNOME/vte/-/issues/253> for more information.
   } deriving (Eq, Generic, FromJSON, Show, ToJSON)
 
 instance FromJSON CursorBlinkMode where
@@ -439,6 +444,7 @@ instance ToJSON CursorBlinkMode where
 --           , showTabBar = ShowTabBarIfNeeded
 --           , cursorBlinkMode = CursorBlinkModeOn
 --           , boldIsBright = False
+--           , enableSixel = False
 --           }
 --   in defaultConfigOptions == defConfOpt
 -- :}
@@ -455,6 +461,7 @@ defaultConfigOptions =
     , showTabBar = ShowTabBarIfNeeded
     , cursorBlinkMode = CursorBlinkModeOn
     , boldIsBright = False
+    , enableSixel = False
     }
 
 -- | The Termonad 'ConfigOptions' along with the 'ConfigHooks'.